Add SEO metadata and restructure pages
